fix RoomService

+++ src/main/java/ru/example/letterflow/service/RoomService.java
@@ -4,7 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.example.letterflow.domain.dto.RoomDto;
+import ru.example.letterflow.domain.dto.UserDto;
 import ru.example.letterflow.domain.entity.Room;
+import ru.example.letterflow.exceptions.InsufficientAccessRightsException;
 import ru.example.letterflow.exceptions.RoomAlreadyExistException;
 import ru.example.letterflow.repository.RoomRepo;
 import ru.example.letterflow.service.mapping.RoomMapper;
@@ -18,11 +20,15 @@ public class RoomService {
     private RoomRepo roomRepo;
 
     @Transactional
-    public RoomDto addRoom(RoomDto roomDto) throws RoomAlreadyExistException {
+    public RoomDto addRoom(RoomDto roomDto, UserDto userDto) throws RoomAlreadyExistException, InsufficientAccessRightsException {
+        if(userDto.isBlocked()){
+            throw new InsufficientAccessRightsException("Заблокированные пользователи не могут создавать комнаты");
+        }
         if(roomRepo.findByRoomName(roomDto.getRoomName()) != null){
             throw new RoomAlreadyExistException("Комната с таким именем уже существует");
         }
         Room room = RoomMapper.ROOM_MAPPER.toEntity(roomDto);
+        room.setUserId(userDto.getUserId());
         roomRepo.save(room);
         return RoomMapper.ROOM_MAPPER.toDto(room);
     }
@@ -39,9 +45,12 @@ public class RoomService {
     }
 
     @Transactional
-    public RoomDto renameRoom(RoomDto roomDto, String string) throws RoomAlreadyExistException {
+    public RoomDto renameRoom(UserDto userDto, RoomDto roomDto, String string) throws RoomAlreadyExistException, InsufficientAccessRightsException {
 //        получить чат из дто или из репозитория?
 //        Room room = RoomMapper.ROOM_MAPPER.toEntity(roomDto);
+        if(userDto.isBlocked()){
+            throw new InsufficientAccessRightsException("Заблокированные пользователи не могут переименовывать комнаты");
+        }
         if(roomRepo.findByRoomName(string) != null){
             throw new RoomAlreadyExistException("Комната с таким именем уже существует");
         }
@@ -51,7 +60,10 @@ public class RoomService {
     }
 
     @Transactional
-    public String deleteRoom(RoomDto roomDto){
+    public String deleteRoom(UserDto userDto, RoomDto roomDto) throws InsufficientAccessRightsException {
+        if(userDto.getUserId() != roomDto.getUserId()){
+            throw new InsufficientAccessRightsException("Чат может удалить только его создатель");
+        }
         roomRepo.deleteById(roomDto.getRoomId());
         return "Чат удален";
     }
